Add fixes for race conditions
We found that the allreduce code has race conditions: The input tensor is not
completely generated before the EnqueueTensorAllreduce code starts, so the
BackgroundThreadLoop appears to copy incorrect data into the beginning of the
allreduce operation. We believe this issue is caused by the streams incorrectly
synchronizing. To mitigate this problem (but probably not fix it!), do not
create separate streams for the BackgroundThreadLoop, but instead, let them
push copies and kernels onto the default stream. This ensures these activities
will serialize with the input tensor's production.

TODO: We need to decide whether to continue debugging this implementation or
to rebuild a different way.